telegram_store_bot/requirements.txt
python-telegram-bot==21.0
aiohttp==3.9.1
qrcode[pil]==7.4.2
Pillow==10.2.0
aiosqlite==0.19.0
python-dotenv==1.0.0
